+++ tests/testthat/test-legend.R
@@ -1,4 +1,9 @@
 testthat::test_that('Symbols', {
+  mapData <- data.frame(x = 1:2,
+                        lng = c(-122, -121),
+                        lat = c(41,42))
+  m <- leaflet::leaflet()
+  pal <- leaflet::colorNumeric('Reds', 1:4)
   # test args
   makeSymbol('notashape', width = 1, color = 'black') %>%
     testthat::expect_error()
@@ -27,6 +32,35 @@ testthat::test_that('Symbols', {
   makeLegendSymbol(shape = 'rect', width = 1, color = 'black') %>%
     testthat::expect_error()
   makeLegendSymbol(label = '', shape = 'rect', width = 1, color = 'black') %>%
+    testthat::expect_error()
+  m %>%
+    addSymbols(lat = ~lat, lng = ~lng, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbols(lat = ~lat, lng = ~lng, color = 'black',
+               values = 1:8, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbols(lat = ~lat, lng = ~lng, color = 'black',
+               shape = c('rect', 'notashape'),
+               values = ~x, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbolsSize(lat = ~lat, lng = ~lng, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbolsSize(lat = ~lat, lng = ~lng,
+                   values = ~x, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbolsSize(lat = ~lat, lng = ~lng, color = 'black',
+               shape = c('notashape'),
+               values = ~x, data = mapData) %>%
+    testthat::expect_error()
+  m %>%
+    addSymbolsSize(lat = ~lat, lng = ~lng, color = 'black',
+                   baseSize = -1,
+                   values = ~x, data = mapData) %>%
     testthat::expect_error()
   # test shapes
   makeSymbol('rect', width = 1, height = 2,
@@ -132,31 +166,28 @@ testthat::test_that('Symbols', {
   sizeBreaks(1:4, 1:4, 5) %>%
     testthat::expect_equal(stats::setNames(c(2, 4, 6, 8), c(1, 2, 3, 4)))
   # test size icons args
-  pal <- leaflet::colorNumeric('Reds', 1:4)
-  makeSymbolsSize(values = 1:4, shape = 'notashape',
-                color = 'black', baseSize = 5) %>%
+  makeSymbolsSize(values = 1:4, shape = 'notashape', fillColor = 'black',
+                  color = 'black', baseSize = 5) %>%
     testthat::expect_error()
-  makeSymbolsSize(values = 1:4, shape = 'rect',
-                color = 'black', baseSize = -5) %>%
+  makeSymbolsSize(values = 1:4, shape = 'rect', fillColor = 'black',
+                  color = 'black', baseSize = -5) %>%
     testthat::expect_error()
-  makeSymbolsSize(values = 1:4, shape = 'rect',
-                color = 'black', baseSize = 5, opacity = '1') %>%
+  makeSymbolsSize(values = 1:4, shape = 'rect', fillColor = 'black',
+                  color = 'black', baseSize = 5, opacity = '1') %>%
     testthat::expect_error()
   makeSymbolsSize(values = 1:4, shape = c('rect', 'circle'),
-                color = 'black', baseSize = 5, opacity = 1) %>%
+                  fillColor = 'black',
+                  color = 'black', baseSize = 5, opacity = 1) %>%
     testthat::expect_error()
-  makeSymbolsSize(values = 1:4, shape = 'rect',
-                color = 'black', baseSize = 5, opacity = 1,
-                strokeWidth = -1) %>%
-    testthat::expect_error()
-  makeSymbolsSize(values = 1:4, shape = 'rect',
-                colorValues = 4:1, baseSize = 5, opacity = 1) %>%
+  makeSymbolsSize(values = 1:4, shape = 'rect', fillColor = 'black',
+                  color = 'black', baseSize = 5, opacity = 1,
+                  strokeWidth = -1) %>%
     testthat::expect_error()
   makeSymbolsSize(values = 1:4, shape = 'rect', baseSize = 5, opacity = 1) %>%
     testthat::expect_error()
   makeSymbolsSize(values = 1:4, shape = 'rect', baseSize = 5, opacity = 1,
-                color = 'black',
-                fillColor = c('black', 'blue')) %>%
+                  color = 'black',
+                  fillColor = c('black', 'blue')) %>%
     testthat::expect_error()
   # test size icons
   makeSymbolsSize(values = 1:4, shape = 'rect', fillColor = 'black',
@@ -190,22 +221,43 @@ testthat::test_that('Symbols', {
   <rect id="rect" x="1" y="1" height="8" width="8" stroke="#67000D" fill="#FFF5F0" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
 </svg>')
   mapData <- data.frame(x = 1:2, lat = c(41,42), lng = c(-122, -122))
-  leaflet::leaflet(mapData) %>%
-    leaflet::addTiles() %>%
-    addSymbols(lng = ~lng, lat = ~lat, width = 20, color = 'black')
-  leaflet::leaflet(mapData) %>%
-    addSymbols(lng = ~lng, lat = ~lat, values = ~x, width = 20, color = 'black')
-  leaflet::leaflet(mapData) %>%
-    addSymbols(lng = ~lng, lat = ~lat, values = ~x, width = 20, color = ~pal(x))
-  leaflet::leaflet(mapData) %>%
-    addSymbols(lng = ~lng, lat = ~lat, values = ~x, width = 20,
-               color = 'black', fillColor = ~pal(x))
+  m %>%
+    addSymbols(lat = ~lat, lng = ~lng, color = ~pal(x), fillColor = ~pal(x),
+               values = ~x, data = mapData) %>%
+    getElement('x') %>%
+    getElement('calls') %>%
+    getElement(1) %>%
+    getElement('args') %>%
+    getElement(3) %>%
+    getElement('iconUrl') %>%
+    getElement('data') %>%
+    URLdecode() %>%
+    testthat::expect_equal(
+      c("data:image/svg+xml,<svg xmlns=\"http://www.w3.org/2000/svg\" version=\"1.1\" width=\"22\" height=\"22\">\n  <rect id=\"rect\" x=\"1\" y=\"1\" height=\"20\" width=\"20\" stroke=\"#FFF5F0\" fill=\"#FFF5F0\" stroke-opacity=\"1\" fill-opacity=\"1\" stroke-width=\"1\"></rect>\n</svg>",
+        "data:image/svg+xml,<svg xmlns=\"http://www.w3.org/2000/svg\" version=\"1.1\" width=\"22\" height=\"22\">\n  <circle id=\"circle\" cx=\"11\" cy=\"11\" r=\"10\" stroke=\"#FDA081\" fill=\"#FDA081\" stroke-opacity=\"1\" fill-opacity=\"1\" stroke-width=\"1\"></circle>\n</svg>"
+      ))
+  m %>%
+    addSymbolsSize(lat = ~lat, lng = ~lng, color = ~pal(x), fillColor = ~pal(x),
+                   values = ~x, data = mapData) %>%
+    getElement('x') %>%
+    getElement('calls') %>%
+    getElement(1) %>%
+    getElement('args') %>%
+    getElement(3) %>%
+    getElement('iconUrl') %>%
+    getElement('data') %>%
+    URLdecode() %>%
+    testthat::expect_equal(
+      c("data:image/svg+xml,<svg xmlns=\"http://www.w3.org/2000/svg\" version=\"1.1\" width=\"15.3333333333333\" height=\"15.3333333333333\">\n  <rect id=\"rect\" x=\"1\" y=\"1\" height=\"13.3333333333333\" width=\"13.3333333333333\" stroke=\"#FFF5F0\" fill=\"#FFF5F0\" stroke-opacity=\"1\" fill-opacity=\"1\" stroke-width=\"1\"></rect>\n</svg>",
+        "data:image/svg+xml,<svg xmlns=\"http://www.w3.org/2000/svg\" version=\"1.1\" width=\"28.6666666666667\" height=\"28.6666666666667\">\n  <rect id=\"rect\" x=\"1\" y=\"1\" height=\"26.6666666666667\" width=\"26.6666666666667\" stroke=\"#FDA081\" fill=\"#FDA081\" stroke-opacity=\"1\" fill-opacity=\"1\" stroke-width=\"1\"></rect>\n</svg>"
+      ))
 
 })
 
-testthat::test_that('Size Legends', {
+testthat::test_that('Symbol Legends', {
   mapData <- data.frame(x = 1:2)
   pal <- leaflet::colorNumeric('Reds', mapData[['x']])
+  factorPal <- leaflet::colorFactor('Set1', mapData[['x']])
   m <- leaflet::leaflet()
   # test args
   m %>%
@@ -233,6 +285,26 @@ testthat::test_that('Size Legends', {
     addLegendLine(data = mapData, values = ~x, color = 'black',
                   numberFormat = 'fun') %>%
     testthat::expect_error()
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x) %>%
+    testthat::expect_error()
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x, pal = factorPal,
+                    fillColor = c('blue', 'red', 'green')) %>%
+    testthat::expect_error()
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x,
+                    color = c('black', 'grey', 'blue'),
+                    pal = factorPal) %>%
+    testthat::expect_error()
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x, color = 'black',
+                    shape = c('rect', 'notashape')) %>%
+    testthat::expect_error()
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x, color = 'black',
+                    shape = c('rect')) %>%
+    testthat::expect_error()
   # test size legends
   m %>%
     addLegendSize(data = mapData, values = ~x, pal = pal, breaks = 1,
@@ -245,15 +317,15 @@ testthat::test_that('Size Legends', {
     URLdecode() %>%
     testthat::expect_equal(
       '<div>
-  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="8.66666666666667" height="8.66666666666667">
-  <rect id="rect" x="1" y="1" height="6.66666666666667" width="6.66666666666667" stroke="black" fill="#FFF5F0" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
-</svg>" style="vertical-align: middle; margin: 5px; margin-right: 3.33333333333333px; margin-left: 3.33333333333333px" height="6.66666666666667" width="6.66666666666667"/>
+  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="15.3333333333333" height="15.3333333333333">
+  <rect id="rect" x="1" y="1" height="13.3333333333333" width="13.3333333333333" stroke="black" fill="#FFF5F0" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
+</svg>" style="vertical-align: middle; margin: 5px; margin-right: 6.66666666666667px; margin-left: 6.66666666666667px" height="13.3333333333333" width="13.3333333333333"/>
   <span style="">1</span>
 </div>
 <div>
-  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="15.3333333333333" height="15.3333333333333">
-  <rect id="rect" x="1" y="1" height="13.3333333333333" width="13.3333333333333" stroke="black" fill="#67000D" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
-</svg>" style="vertical-align: middle; margin: 5px; margin-right: 0px; margin-left: 0px" height="13.3333333333333" width="13.3333333333333"/>
+  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="28.6666666666667" height="28.6666666666667">
+  <rect id="rect" x="1" y="1" height="26.6666666666667" width="26.6666666666667" stroke="black" fill="#67000D" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
+</svg>" style="vertical-align: middle; margin: 5px; margin-right: 0px; margin-left: 0px" height="26.6666666666667" width="26.6666666666667"/>
   <span style="">2</span>
 </div>')
   # test line legends
@@ -278,6 +350,29 @@ testthat::test_that('Size Legends', {
 </svg>" style="vertical-align: middle; margin: 5px; margin-right: 0px; margin-left: 0px" height="13.3333333333333" width="20"/>
   <span style="">2</span>
 </div>')
+  m %>%
+    addLegendSymbol(data = mapData, values = ~x, color = 'black',
+                    pal = factorPal) %>%
+    getElement(1) %>%
+    getElement('calls') %>%
+    getElement(1) %>%
+    getElement('args') %>%
+    getElement(1) %>%
+    URLdecode() %>%
+    testthat::expect_equal(
+      '<div>
+  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="22" height="22">
+  <rect id="rect" x="1" y="1" height="20" width="20" stroke="black" fill="#E41A1C" stroke-opacity="1" fill-opacity="1" stroke-width="1"></rect>
+</svg>" style="vertical-align: middle; margin: 5px; margin-right: 0px; margin-left: 0px" height="20" width="20"/>
+  <span style="">1</span>
+</div>
+<div>
+  <img src="data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="22" height="22">
+  <circle id="circle" cx="11" cy="11" r="10" stroke="black" fill="#4DAF4A" stroke-opacity="1" fill-opacity="1" stroke-width="1"></circle>
+</svg>" style="vertical-align: middle; margin: 5px; margin-right: 0px; margin-left: 0px" height="20" width="20"/>
+  <span style="">2</span>
+</div>')
+
 })
 
 testthat::test_that('Image Legend', {
@@ -461,7 +556,7 @@ testthat::test_that('Categorical Legends', {
                         y = rep(LETTERS[1:2], 5))
   m <- leaflet::leaflet(mapData)
   palQuantile <- leaflet::colorQuantile('Reds', mapData[['x']], n = 2)
-  palBins <- leaflet::colorBin('Blues', mapData[['x']], bins = 2)
+  palBin <- leaflet::colorBin('Blues', mapData[['x']], bins = 2)
   palFactor <- leaflet::colorFactor('viridis', mapData[['y']])
   # test Quantile args
   m %>% addLegendQuantile(pal = palQuantile,
@@ -490,27 +585,21 @@ testthat::test_that('Categorical Legends', {
     testthat::expect_error()
   # test Bin args
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      orientation = 'up') %>%
     testthat::expect_error()
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      title = 1:2) %>%
     testthat::expect_error()
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      shape = 'notashape') %>%
     testthat::expect_error()
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      width = -1) %>%
     testthat::expect_error()
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      height = -1) %>%
     testthat::expect_error()
   m %>% addLegendBin(pal = palBin,
-                     values = ~x,
                      numberFormat = 'fun') %>%
     testthat::expect_error()
   # test Factor args
@@ -568,8 +657,7 @@ testthat::test_that('Categorical Legends', {
 </svg>" style="vertical-align: middle; padding: 1px;"/>
 <span style="vertical-align: middle; padding: 1px; "> 50% - 100% (6 - 10)</span>
 <br/>')
-  m %>% addLegendBin(pal = palBins,
-                     values = ~x) %>%
+  m %>% addLegendBin(pal = palBin) %>%
     getElement(1) %>%
     getElement('calls') %>%
     getElement(1) %>%
@@ -692,3 +780,4 @@ testthat::test_that('Helper Functions', {
   leafletAmBootstrapDependencies() %>%
     testthat::expect_equal(leaflet:::leafletAmBootstrapDependencies())
 })
+